refactor(components): Shifting directory structure
