+++ src/components/holidays/omer/utilities/getOmer.spec.ts
@@ -1,5 +1,5 @@
-import { local } from '../../../../test/jest-framework';
-import { IHoliday } from '../../../holidays/types';
+import { local } from '../../../../../test/jest-framework';
+import { IHoliday } from '../../types';
 import getOmer, { blessings, getOmerBlessingInfo } from './getOmer';
 
 describe('omer logic', () => {
