+++ src/screens/HomeScreen.tsx
@@ -5,8 +5,8 @@ import { styles } from 'is-it-shabbat-core';
 import React from 'react';
 import { Dimensions, ScrollView, StatusBar, StyleProp, View, ViewStyle } from 'react-native';
 
-import Holidays from '../components/Holidays';
-import IsItShabbat from '../components/IsItShabbat';
+import Holidays from '../components/holidays/Holidays';
+import IsItShabbat from '../components/shabbat/IsItShabbat';
 import { BackgroundView, CenteredContainer, Footer, SubtitleCenterText } from '../Styles';
 
 const getBannerSize = ({ width }: { width: number }): 'fullBanner' | 'largeBanner' => {
