+++ src/screens/OmerScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Omer from '../components/Omer';
+import Omer from '../components/holidays/omer/Omer';
 import { BackgroundView } from '../Styles';
 
 export default () => (
