+++ src/screens/SettingsScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import LocationPicker from '../components/LocationPicker';
-import UseCurrentLocationButton from '../components/UseCurrentLocationButton';
+import LocationPicker from '../components/location/LocationPicker';
+import UseCurrentLocationButton from '../components/location/UseCurrentLocationButton';
 import { BackgroundView } from '../Styles';
 
 export default () => (
